# models/song.go
package models

import "time"

// Song описание песни
// @Description Структура песни
// @Accept json
// @Produce json

type Song struct {
	ID          int       `json:"id"`
	Group       string    `json:"group"`
	Song        string    `json:"song"`
	ReleaseDate time.Time `json:"release_date"`
	Text        string    `json:"text"`
	Link        string    `json:"link"`
}